requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==18.2.0
cffi==1.12.2
chardet==3.0.4
discord.py==1.0.0a1696+g48b60b2
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
pkg-resources==0.0.0
pycparser==2.19
PyNaCl==1.2.1
six==1.12.0
typing-extensions==3.7.2
websockets==6.0
yarl==1.3.0
